add semantic versioning to build script; resolves #51

+++ module.build.ps1
@@ -1,5 +1,6 @@
-
+#requires -Modules InvokeBuild, PSDeploy, BuildHelpers, PSScriptAnalyzer, PlatyPS, Pester
 $script:ModuleName = 'PSGraph'
+
 $script:Source = Join-Path $BuildRoot $ModuleName
 $script:Output = Join-Path $BuildRoot output
 $script:Destination = Join-Path $Output $ModuleName
@@ -7,14 +8,47 @@ $script:ModulePath = "$Destination\$ModuleName.psm1"
 $script:ManifestPath = "$Destination\$ModuleName.psd1"
 $script:Imports = ( 'private', 'public', 'classes' )
 $script:TestFile = "$PSScriptRoot\output\TestResults_PS$PSVersion`_$TimeStamp.xml"
+$script:HelpRoot = Join-Path $Output 'help'
 
-Task Default Build, Pester, UpdateSource, Publish
+function ComplexTask
+{
+    [CmdletBinding()]
+    param(
+        [Parameter(Position = 0, Mandatory = 1)]
+        [string]
+        $Name,
+        [Parameter(Position = 1, Mandatory = 1)]
+        [hashtable]
+        $Options
+    )
+    try     
+    {
+        if ( $null -ne $Options.Action )
+        {
+            $Options.Jobs = $Options.Action
+            $Options.Remove('Action')
+        }
+        Task $Name @Options
+    }
+    catch 
+    {
+        Write-Error "Failed to execute taks [$Name]"
+        throw
+    }    
+}
+
+Task Default Clean, Build, Pester, UpdateSource, Publish
 Task Build CopyToOutput, BuildPSM1, BuildPSD1
 Task Pester Build, ImportModule, UnitTests, FullTests
+Task Local Build, Pester, UpdateSource
 
 Task Clean {
-    $null = Remove-Item $Output -Recurse -ErrorAction Ignore
-    $null = New-Item  -Type Directory -Path $Destination
+    
+    If (Test-Path $Output)
+    {
+        $null = Remove-Item $Output -Recurse -ErrorAction Ignore
+    }
+    $null = New-Item  -Type Directory -Path $Destination -ErrorAction Ignore
 }
 
 Task UnitTests {
@@ -58,62 +92,109 @@ Task CopyToOutput {
         ForEach-Object { "  Create [.{0}]" -f $_.fullname.replace($PSScriptRoot, '')}
 }
 
-Task BuildPSM1 -Inputs (Get-Item "$source\*\*.ps1") -Outputs $ModulePath {
-
-    [System.Text.StringBuilder]$stringbuilder = [System.Text.StringBuilder]::new()
-    foreach ($folder in $imports )
-    {
-        [void]$stringbuilder.AppendLine( "Write-Verbose 'Importing from [$Source\$folder]'" )
-        if (Test-Path "$source\$folder")
+ComplexTask BuildPSM1 @{
+    Inputs  = (Get-Item "$source\*\*.ps1") 
+    Outputs = $ModulePath 
+    Action  = {
+        [System.Text.StringBuilder]$stringbuilder = [System.Text.StringBuilder]::new()    
+        foreach ($folder in $imports )
         {
-            $fileList = Get-ChildItem "$source\$folder\*.ps1" | Where Name -NotLike '*.Tests.ps1'
-            foreach ($file in $fileList)
+            [void]$stringbuilder.AppendLine( "Write-Verbose 'Importing from [$Source\$folder]'" )
+            if (Test-Path "$source\$folder")
             {
-                $shortName = $file.fullname.replace($PSScriptRoot, '')
-                "  Importing [.$shortName]"
-                [void]$stringbuilder.AppendLine( "# .$shortName" )
-                [void]$stringbuilder.AppendLine( [System.IO.File]::ReadAllText($file.fullname) )
+                $fileList = Get-ChildItem "$source\$folder\*.ps1" | Where Name -NotLike '*.Tests.ps1'
+                foreach ($file in $fileList)
+                {
+                    $shortName = $file.fullname.replace($PSScriptRoot, '')
+                    "  Importing [.$shortName]"
+                    [void]$stringbuilder.AppendLine( "# .$shortName" ) 
+                    [void]$stringbuilder.AppendLine( [System.IO.File]::ReadAllText($file.fullname) )
+                }
             }
         }
+        
+        "  Creating module [$ModulePath]"
+        Set-Content -Path  $ModulePath -Value $stringbuilder.ToString() 
     }
-
-    "  Creating module [$ModulePath]"
-    Set-Content -Path  $ModulePath -Value $stringbuilder.ToString()
 }
 
-Task NextPSGalleryVersion -if (-Not ( Test-Path "$output\version.xml" ) ) -Before BuildPSD1 {
-    $galleryVersion = Get-NextPSGalleryVersion -Name $ModuleName
-    $galleryVersion | Export-Clixml -Path "$output\version.xml"
+ComplexTask NextPSGalleryVersion @{
+    If     = (-Not ( Test-Path "$output\version.xml" ) ) 
+    Before = 'BuildPSD1'
+    Action = {
+        $galleryVersion = Get-NextPSGalleryVersion -Name $ModuleName
+        $galleryVersion | Export-Clixml -Path "$output\version.xml"
+    }
 }
 
-Task BuildPSD1 -inputs (Get-ChildItem $Source -Recurse -File) -Outputs $ManifestPath {
+ComplexTask BuildPSD1 @{
+    Inputs  = (Get-ChildItem $Source -Recurse -File) 
+    Outputs = $ManifestPath 
+    Action  = {
+    
+        Write-Output "  Update [$ManifestPath]"
+        Copy-Item "$source\$ModuleName.psd1" -Destination $ManifestPath
 
-    "  Update [$ManifestPath]"
-    Copy-Item "$source\$ModuleName.psd1" -Destination $ManifestPath
 
-    $bumpVersionType = 'Patch'
+        $functions = Get-ChildItem "$ModuleName\Public\*.ps1" | Where-Object { $_.name -notmatch 'Tests'} | Select-Object -ExpandProperty basename      
+        Set-ModuleFunctions -Name $ManifestPath -FunctionsToExport $functions
 
-    $functions = Get-ChildItem "$ModuleName\Public\*.ps1" | Where-Object { $_.name -notmatch 'Tests'} | Select-Object -ExpandProperty basename
+        Write-Output "  Detecting semantic versioning"
 
-    $oldFunctions = (Get-Metadata -Path $manifestPath -PropertyName 'FunctionsToExport')
+        Import-Module ".\$ModuleName"
+        $commandList = Get-Command -Module $ModuleName
+        Remove-Module $ModuleName
 
-    $functions | Where {$_ -notin $oldFunctions } | % {$bumpVersionType = 'Minor'}
-    $oldFunctions | Where {$_ -notin $Functions } | % {$bumpVersionType = 'Major'}
+        Write-Output "    Calculating fingerprint"
+        $fingerprint = foreach ($command in $commandList )
+        {
+            foreach ($parameter in $command.parameters.keys)
+            {
+                '{0}:{1}' -f $command.name, $command.parameters[$parameter].Name
+                $command.parameters[$parameter].aliases | Foreach-Object { '{0}:{1}' -f $command.name, $_}
+            }
+        }
+        
+        if (Test-Path .\fingerprint)
+        {
+            $oldFingerprint = Get-Content .\fingerprint
+        }
+        
+        $bumpVersionType = 'Patch'
+        '    Detecting new features'
+        $fingerprint | Where {$_ -notin $oldFingerprint } | % {$bumpVersionType = 'Minor'; "      $_"}    
+        '    Detecting breaking changes'
+        $oldFingerprint | Where {$_ -notin $fingerprint } | % {$bumpVersionType = 'Major'; "      $_"}
 
-    Set-ModuleFunctions -Name $ManifestPath -FunctionsToExport $functions
+        Set-Content -Path .\fingerprint -Value $fingerprint
 
-    # Bump the module version
-    $version = [version] (Get-Metadata -Path $manifestPath -PropertyName 'ModuleVersion')
-    $galleryVersion = Import-Clixml -Path "$output\version.xml"
-    if ( $version -lt $galleryVersion )
-    {
-        $version = $galleryVersion
+        # Bump the module version
+        $version = [version] (Get-Metadata -Path $manifestPath -PropertyName 'ModuleVersion')
+
+        if ( $version -lt ([version]'1.0.0') )
+        {
+            # Still in beta, don't bump major version
+            if ( $bumpVersionType -eq 'Major'  )
+            {
+                $bumpVersionType = 'Minor'
+            }
+            else 
+            {
+                $bumpVersionType = 'Patch'
+            }       
+        }
+
+        $galleryVersion = Import-Clixml -Path "$output\version.xml"
+        if ( $version -lt $galleryVersion )
+        {
+            $version = $galleryVersion
+        }
+        Write-Output "  Stepping [$bumpVersionType] version [$version]"
+        $version = [version] (Step-Version $version -Type $bumpVersionType)
+        Write-Output "  Using version: $version"
+        
+        Update-Metadata -Path $ManifestPath -PropertyName ModuleVersion -Value $version
     }
-    "  Stepping [$bumpVersionType] version [$version]"
-    $version = [version] (Step-Version $version -Type $bumpVersionType)
-    "  Using version: $version"
-
-    Update-Metadata -Path $ManifestPath -PropertyName ModuleVersion -Value $version
 }
 
 Task UpdateSource {
@@ -159,5 +240,39 @@ Task Publish {
         "`t* You are in a known build system (Current: $ENV:BHBuildSystem)`n" +
         "`t* You are committing to the master branch (Current: $ENV:BHBranchName) `n" +
         "`t* Your commit message includes !deploy (Current: $ENV:BHCommitMessage)"
+    }
+}
+
+ComplexTask CreateHelp @{
+    Inputs  = {Get-ChildItem "$ModuleName\Public\*.ps1"}
+    Outputs = {
+        process
+        {
+            Get-ChildItem $_ | % {'{0}\{1}.md' -f $HelpRoot, $_.basename}
+        }
+    }
+    Partial = $true
+    Action  = 'ImportModule', {    
+        process
+        {
+            $null = New-Item -Path $HelpRoot -ItemType Directory -ErrorAction SilentlyContinue        
+            $mdHelp = @{
+                #Module                = $script:ModuleName
+                OutputFolder          = $HelpRoot
+                AlphabeticParamsOrder = $true
+                Verbose               = $true
+                Force                 = $true
+                Command               = Get-Item $_ | % basename
+            }
+            New-MarkdownHelp @mdHelp | % fullname
+        }    
+    }
+}
+
+ComplexTask PackageHelp  @{
+    Inputs  = {Get-ChildItem $HelpRoot -Recurse -File}
+    Outputs = "$Destination\en-us\$ModuleName-help.xml"
+    Action  = 'CreateHelp', {
+        New-ExternalHelp -Path $HelpRoot -OutputPath "$Destination\en-us" -force | % fullname
     }
 }
